1.FinAssist_chatbot/requirements.txt
streamlit
transformers
torch
torchvision
torchaudio
accelerate
python-dotenv
plotly
pandas
numpy
matplotlib
seaborn
streamlit-option-menu
streamlit-aggrid
streamlit-authenticator
pymongo
pymongo[srv]
python-multipart
soundfile
librosa
gradio
streamlit-extras
streamlit-echarts
streamlit-plotly-events
SpeechRecognition
gtts
pyaudio
